% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-paths.R
\name{notebook_dir}
\alias{notebook_dir}
\title{Return the notebook directory}
\usage{
notebook_dir(project_dir = getwd(), folder = "notebook")
}
\arguments{
\item{project_dir}{Project directory. Defaults to \code{getwd()}.}

\item{folder}{Name of the notebook folder.}
}
\value{
A character path.
}
\description{
Constructs the path to the project notebook directory. By default,
\code{project_dir} is the current working directory.
}
\examples{
\dontrun{
notebook_dir()
notebook_dir(project_dir = "/path/to/project")
notebook_dir(folder = "analysis_notes")
}
}
