% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-display.R
\name{show_done_tasks}
\alias{show_done_tasks}
\title{Show completed notebook tasks}
\usage{
show_done_tasks(note_dir = notebook_dir())
}
\arguments{
\item{note_dir}{Directory containing the notebook files.}
}
\value{
The completed tasks, invisibly.
}
\description{
Prints all tasks that have been marked as completed.
}
\examples{
\dontrun{
show_done_tasks()
}
}
