% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-display.R
\name{show_general_notes}
\alias{show_general_notes}
\title{Show general notebook notes}
\usage{
show_general_notes(note_dir = notebook_dir())
}
\arguments{
\item{note_dir}{Directory containing the notebook files.}
}
\value{
The general notes, invisibly.
}
\description{
Prints all general notebook notes.
}
\examples{
\dontrun{
show_general_notes()
}
}
