% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-display.R
\name{show_notes}
\alias{show_notes}
\title{Show open notebook notes}
\usage{
show_notes(note_dir = notebook_dir())
}
\arguments{
\item{note_dir}{Directory containing the notebook files.}
}
\value{
The open notes, invisibly.
}
\description{
Prints all currently open notebook notes.
}
\examples{
\dontrun{
show_notes()
}
}
