% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-display.R
\name{show_tasks}
\alias{show_tasks}
\title{Show open notebook tasks}
\usage{
show_tasks(note_dir = notebook_dir())
}
\arguments{
\item{note_dir}{Directory containing the notebook files.}
}
\value{
The open tasks, invisibly.
}
\description{
Prints all currently open notebook tasks.
}
\examples{
\dontrun{
show_tasks()
}
}
